# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Hemansh Manager API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List all tasks
      parameters:
        - name: category
          in: query
          schema:
            type: string
            enum: [Study, Gym, Personal]
        - name: priority
          in: query
          schema:
            type: string
            enum: [High, Medium, Low]
        - name: completed
          in: query
          schema:
            type: boolean
        - name: date
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskBody"
      responses:
        "201":
          description: Created task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    get:
      operationId: getTask
      tags: [tasks]
      summary: Get a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskBody"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── STUDY ─────────────────────────────────────────────────────────────
  /study/sessions:
    get:
      operationId: listStudySessions
      tags: [study]
      summary: List study sessions
      parameters:
        - name: date
          in: query
          schema:
            type: string
            format: date
        - name: week
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of study sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StudySession"
    post:
      operationId: createStudySession
      tags: [study]
      summary: Log a study session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStudySessionBody"
      responses:
        "201":
          description: Created study session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudySession"
  /study/sessions/{id}:
    patch:
      operationId: updateStudySession
      tags: [study]
      summary: Update a study session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStudySessionBody"
      responses:
        "200":
          description: Updated session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudySession"
    delete:
      operationId: deleteStudySession
      tags: [study]
      summary: Delete a study session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Task:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        category:
          type: string
          enum: [Study, Gym, Personal]
        priority:
          type: string
          enum: [High, Medium, Low]
        completed:
          type: boolean
        deadline:
          type: string
          format: date-time
          nullable: true
        date:
          type: string
          format: date
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, title, category, priority, completed, createdAt, updatedAt]
    CreateTaskBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        category:
          type: string
          enum: [Study, Gym, Personal]
        priority:
          type: string
          enum: [High, Medium, Low]
        deadline:
          type: string
          format: date-time
          nullable: true
        date:
          type: string
          format: date
          nullable: true
      required: [title, category, priority]
    UpdateTaskBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        category:
          type: string
          enum: [Study, Gym, Personal]
        priority:
          type: string
          enum: [High, Medium, Low]
        completed:
          type: boolean
        deadline:
          type: string
          format: date-time
          nullable: true
        date:
          type: string
          format: date
          nullable: true

    # ── Study ──
    StudySession:
      type: object
      properties:
        id:
          type: integer
        subject:
          type: string
        hours:
          type: number
        date:
          type: string
          format: date
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, subject, hours, date, createdAt]
    CreateStudySessionBody:
      type: object
      properties:
        subject:
          type: string
        hours:
          type: number
        date:
          type: string
          format: date
        notes:
          type: string
          nullable: true
      required: [subject, hours, date]
    UpdateStudySessionBody:
      type: object
      properties:
        subject:
          type: string
        hours:
          type: number
        date:
          type: string
          format: date
        notes:
          type: string
          nullable: true
...