# lib/api-zod/src/generated/types/createStudySessionBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hemansh Manager API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateStudySessionBody {
  subject: string;
  hours: number;
  date: Date;
  notes?: string | null;
}