# lib/api-zod/src/generated/types/updateTaskBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hemansh Manager API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateTaskBodyCategory } from "./updateTaskBodyCategory";
import type { UpdateTaskBodyPriority } from "./updateTaskBodyPriority";

export interface UpdateTaskBody {
  title?: string;
  description?: string | null;
  category?: UpdateTaskBodyCategory;
  priority?: UpdateTaskBodyPriority;
  completed?: boolean;
  deadline?: Date | null;
  date?: Date | null;
}